requirements.txt
aiocache
alembic
anyio
argon2-cffi
asyncpg
bcrypt
build
databases[asyncpg]
faiss-cpu
fastapi[all]
fuzzywuzzy
httpx
huggingface-hub==0.32.0
langchain-huggingface==0.2.0
langchain-core==0.3.61
langchain-community==0.3.0
langchain-text-splitters==0.3.0
mypy
pika
pre-commit
psycopg2
pydantic
"passlib[bcrypt]"
pydantic-settings
pytest
pytest-alembic
pytest-asyncio
pytest-cov
pytest-mock
pytest-mock-resources[docker]
python-dotenv
python-jose
python-Levenshtein
python-multipart
pylint
pytest
pytest-asyncio
pytest-cov
requests==2.32.3
scikit-learn
sentence-transformers
spacy
scispacy
sqlalchemy
sqlmodel
torch
twine
uvicorn
uvicorn[standard]
